added model interface and unit tests
added process interface and tests

added combined interface

added differential cross section interface for abstract processes and models, as well as included tests of the interface

replaced the hard and soft interfaces by just the registration of the function name. Added the respective doc string to indicate the implementations to be for the interfaces.

added doc strings to the functions of the process and model interface. Added an section to the README.md on how-to build the documentation locally.