[WIP] Pretty printing (#60)
Adds pretty printing for:
- PerturbativeQED model
- Compton process
- Coordinate systems, frames of reference, and phase space definition
- Particle stateful
- PhaseSpacePoint

Essentially, adding similar functionality to
https://github.com/QEDjl-project/QEDbase.jl/pull/61 for the QEDprocesses
types.

I'm using the jldoctests to test, I'm not sure if it's really necessary
to test in a separate test as well. Putting single jldoctests for these
prints in every type definition would clutter the documentation a bit so
I've not done that. If we really want tests for each `show()` we should
probably do it in a test file.

Also, I'm not sure if there's a better way for the print functions than
alternating the `show` for objects and `print` for strings.
Interpolating doesn't work unless we also overload the `print`. (I
*think* it is print, the number of string output related functions in
Julia is confusing)