Moved test deps to main Project.toml (#62)
As the title says, this moves the test dependencies from
`test/Project.toml` to the main `Project.toml`. To not mix these with
the package dependencies, we use the `[extras]` and `[target]` blocks.

This keeps the dependencies of the package incl. tests in one place and
makes maintenance easier. Furthermore, it allows testing locally against
unreleased branches of a package dependency, because the package and the
tests use the same `Manifest.toml`.

Co-authored-by: Uwe Hernandez Acosta <u.hernandez@hzdr.de>