Proposal for PhaseSpacePoint improvement (#63)
Proposal for a rework of the PhaseSpacePoint, according to Issue #58

I did a lot of type magic with recursive variadic templates to find out
type information and get functions perfectly type stable. Constructing
phase space points is now always type stable and takes a maximum of
~11ns for me when constructing from momenta.

Some of the tests are failing for now because some of the interfaces
currently don't expect tuples from the PSP implementation. It might make
sense to fix this in #59

---------

Co-authored-by: Uwe Hernandez Acosta <u.hernandez@hzdr.de>
Co-authored-by: Uwe Hernandez Acosta <szabo137@users.noreply.github.com>
Co-authored-by: AntonReinhard <anton.reinhard@protonmail.com>