GPU support (#69)
This PR adds GPU support. It adds tests that run conditionally on
AMDGPU.jl and/or CUDA.jl, depending on whether they are functional on
the machine we're running on.

This PR depends on the fix
https://github.com/QEDjl-project/QEDbase.jl/pull/64 and is rebased to
#68.

Left to do:
- [x] Fix the Compton `_total_probability` function which currently does
not work on GPU because of `quadgk` -> opened issue
https://github.com/QEDjl-project/QEDprocesses.jl/issues/119
- [x] Add remaining tests for the PSP interface on GPU
- [x] Potentially add testing with KernelAbstractions.jl (does that
support broadcasting?), oneAPI.jl and Metal.jl -> Added oneAPI and
Metal, KernelAbstractions would make more sense for actual kernels which
we don't have yet
- [x] Automatic testing in CI with GPU capable runners
- [x] Add version checks for the GPU tests, since not all libraries work
on all versions and there's nothing really we can do about that other
than disabling the tests on those versions -> the checks when deciding
what GPU tests to run should be enough for now
- [x] Use PkgExtensions to only optionally load the GPU libraries, since
this takes a long time and should only be done when specifically
requested.

---------

Co-authored-by: AntonReinhard <anton.reinhard@protonmail.com>