Run fewer GPU tests in CI (#124)
Adds a testing util function `LARGE_TESTS()` that returns true if larger
tests should be run.

If the environment variable `LARGE_TESTS` is explicitly set, it will run
them ("1") or skip them ("0"). If it's not set, they will by default be
run locally and skipped in the CI, i.e., when `CI` is set.

Currently, this just disables some of the spin/pol combinations tested
in the GPU tests, but could be used for more future tests as well.
Currently, there is no way to activate the LARGE_TESTS in the CI for a
specific run, because we cannot manually set environment variables in
the scripts.