KernelAbstractions kernels (#146)
This adds the other side of the interface to use the generated kernels
from QEDFeynmanDiagrams.jl

It also adds tests for the generic QEDbase.jl KernelAbstractions kernels
for the specific 1-photon Compton process, and a new set of tests for
`ScatteringProcess`.

The implemented kernels are the unsafe `diff_prob` and `diff_cs`
variants.